Refactor: Extract storage and processing logic into separate classes
