+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/InMemoryUserStorage.java
@@ -1,0 +1,80 @@
+package ru.yandex.practicum.filmorate.storage.user;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
+
+@Slf4j
+@Component
+public class InMemoryUserStorage implements UserStorage {
+    private final Map<Long, User> users = new HashMap<>();
+
+    @Override
+    public Collection<User> findAll() {
+        return users.values();
+    }
+
+    @Override
+    public User create(User user) {
+        if (users.values().stream().anyMatch(u -> u.getEmail().equals(user.getEmail()))) {
+            String msg = "This email " + user.getEmail() + " is already used";
+            log.error(msg);
+            throw new ValidationException(msg);
+        }
+
+        if (users.values().stream().anyMatch(u -> u.getLogin().equals(user.getLogin()))) {
+            String msg = "Login " + user.getLogin() + " is already used";
+            log.error(msg);
+            throw new ValidationException(msg);
+        }
+
+        if (user.getName() == null || user.getName().isBlank()) {
+            user.setName(user.getLogin());
+        }
+
+        user.setId(getNextId());
+        users.put(user.getId(), user);
+        log.info("User has been added: {}", user);
+        return user;
+    }
+
+    @Override
+    public User updateUser(User newUser) {
+        if (newUser.getId() == null) {
+            String msg = "ID must be specified.";
+            log.error(msg);
+            throw new ValidationException(msg);
+        }
+
+        if (users.containsKey(newUser.getId())) {
+            User oldUser = users.get(newUser.getId());
+            oldUser.setEmail(newUser.getEmail());
+            oldUser.setLogin(newUser.getLogin());
+            if (newUser.getName() == null || newUser.getName().isBlank()) {
+                oldUser.setName(newUser.getLogin());
+            } else {
+                oldUser.setName(newUser.getName());
+            }
+            oldUser.setBirthday(newUser.getBirthday());
+            log.info("User with id = {} has been updated", newUser.getId());
+            return oldUser;
+        }
+        String msg = "User with id  = " + newUser.getId() + " not found.";
+        log.error(msg);
+        throw new ValidationException(msg);
+    }
+
+    private long getNextId() {
+        long currentMaxId = users.keySet()
+                .stream()
+                .mapToLong(id -> id)
+                .max()
+                .orElse(0);
+        return ++currentMaxId;
+    }
+}
